Выполненное домашнее задание №1
