+++ app/src/main/res/drawable/ic_music.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="21dp"
+    android:height="20dp"
+    android:viewportWidth="21"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M18.5,0H6.5C5.4,0 4.5,0.9 4.5,2V14C4.5,15.1 5.4,16 6.5,16H18.5C19.6,16 20.5,15.1 20.5,14V2C20.5,0.9 19.6,0 18.5,0ZM16.5,5H13.5V10.5C13.5,11.88 12.38,13 11,13C9.62,13 8.5,11.88 8.5,10.5C8.5,9.12 9.62,8 11,8C11.57,8 12.08,8.19 12.5,8.51V3H16.5V5ZM2.5,4H0.5V18C0.5,19.1 1.4,20 2.5,20H16.5V18H2.5V4Z"
+      android:fillColor="#1A1B22"/>
+</vector>
