+++ app/src/main/res/drawable/ic_search.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="19dp"
+    android:height="19dp"
+    android:viewportWidth="19"
+    android:viewportHeight="19">
+  <path
+      android:pathData="M13.579,11.95H12.721L12.417,11.656C13.481,10.418 14.122,8.81 14.122,7.061C14.122,3.161 10.961,0 7.061,0C3.161,0 0,3.161 0,7.061C0,10.961 3.161,14.122 7.061,14.122C8.81,14.122 10.418,13.481 11.656,12.417L11.95,12.721V13.579L17.381,19L19,17.381L13.579,11.95ZM7.061,11.95C4.356,11.95 2.173,9.766 2.173,7.061C2.173,4.356 4.356,2.173 7.061,2.173C9.766,2.173 11.95,4.356 11.95,7.061C11.95,9.766 9.766,11.95 7.061,11.95Z"
+      android:fillColor="#1A1B22"/>
+</vector>
