+++ app/src/main/res/drawable/ic_setting.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="20dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M17.14,10.94C17.18,10.64 17.2,10.33 17.2,10C17.2,9.68 17.18,9.36 17.13,9.06L19.16,7.48C19.34,7.34 19.39,7.07 19.28,6.87L17.36,3.55C17.24,3.33 16.99,3.26 16.77,3.33L14.38,4.29C13.88,3.91 13.35,3.59 12.76,3.35L12.4,0.81C12.36,0.57 12.16,0.4 11.92,0.4H8.08C7.84,0.4 7.65,0.57 7.61,0.81L7.25,3.35C6.66,3.59 6.12,3.92 5.63,4.29L3.24,3.33C3.02,3.25 2.77,3.33 2.65,3.55L0.74,6.87C0.62,7.08 0.66,7.34 0.86,7.48L2.89,9.06C2.84,9.36 2.8,9.69 2.8,10C2.8,10.31 2.82,10.64 2.87,10.94L0.84,12.52C0.66,12.66 0.61,12.93 0.72,13.13L2.64,16.45C2.76,16.67 3.01,16.74 3.23,16.67L5.62,15.71C6.12,16.09 6.65,16.41 7.24,16.65L7.6,19.19C7.65,19.43 7.84,19.6 8.08,19.6H11.92C12.16,19.6 12.36,19.43 12.39,19.19L12.75,16.65C13.34,16.41 13.88,16.09 14.37,15.71L16.76,16.67C16.98,16.75 17.23,16.67 17.35,16.45L19.27,13.13C19.39,12.91 19.34,12.66 19.15,12.52L17.14,10.94ZM10,13.6C8.02,13.6 6.4,11.98 6.4,10C6.4,8.02 8.02,6.4 10,6.4C11.98,6.4 13.6,8.02 13.6,10C13.6,11.98 11.98,13.6 10,13.6Z"
+      android:fillColor="#1A1B22"/>
+</vector>
