+++ app/src/main/res/layout/activity_setting.xml
@@ -1,0 +1,144 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main"
+    android:orientation="vertical"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/white"
+    tools:context=".SettingActivity">
+
+    <!--    Используемые иллюстрации (иконки) в образовательных целях предоставлены Яндекс-Практикумом (practicum.yandex.ru)-->
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/dimen_56"
+        android:padding="@dimen/dimen_4">
+
+        <ImageView
+            android:id="@+id/button_back"
+            android:layout_width="@dimen/dimen_48"
+            android:layout_height="@dimen/dimen_48"
+            app:srcCompat="@drawable/ic_back"
+            app:tint="@color/grey_100" />
+
+        <TextView
+            android:id="@+id/panel_header"
+            android:layout_width="wrap_content"
+            android:layout_height="@dimen/dimen_48"
+            android:layout_marginStart="@dimen/dimen_60"
+            android:textColor="@color/grey_100"
+            android:textSize="@dimen/dimen_22s"
+            android:textStyle="bold"
+            android:gravity="center_vertical"
+            android:text="@string/setting" />
+    </FrameLayout>
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/dimen_61">
+
+        <TextView
+            android:id="@+id/header_theme"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@color/black"
+            android:textSize="@dimen/dimen_16s"
+            android:paddingStart="@dimen/dimen_16"
+            android:gravity="center_vertical"
+            android:layout_gravity="start|center_vertical"
+            android:text="@string/dark_theme"
+            tools:ignore="RtlSymmetry" />
+
+        <Switch
+            android:id="@+id/switch_theme"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="end|center_vertical"
+            android:padding="@dimen/dimen_9"
+            android:layout_marginEnd="@dimen/dimen_4"
+            android:thumbTint="@color/grey_500"
+            android:trackTint="@color/grey_700"
+            android:checked="false"
+            tools:ignore="UseSwitchCompatOrMaterialXml" />
+    </FrameLayout>
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/dimen_61">
+
+        <TextView
+            android:id="@+id/header_share"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@color/black"
+            android:textSize="@dimen/dimen_16s"
+            android:paddingStart="@dimen/dimen_16"
+            android:gravity="center_vertical"
+            android:layout_gravity="start|center_vertical"
+            android:text="@string/header_share"
+            tools:ignore="RtlSymmetry" />
+
+        <ImageView
+            android:id="@+id/button_share"
+            android:layout_width="@dimen/dimen_24"
+            android:layout_height="@dimen/dimen_24"
+            app:srcCompat="@drawable/ic_share"
+            app:tint="@color/grey_500"
+            android:layout_gravity="end|center_vertical"
+            android:layout_marginEnd="@dimen/dimen_16"/>
+    </FrameLayout>
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/dimen_61">
+
+        <TextView
+            android:id="@+id/header_support"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@color/black"
+            android:textSize="@dimen/dimen_16s"
+            android:paddingStart="@dimen/dimen_16"
+            android:gravity="center_vertical"
+            android:layout_gravity="start|center_vertical"
+            android:text="@string/header_support"
+            tools:ignore="RtlSymmetry" />
+
+        <ImageView
+            android:id="@+id/button_support"
+            android:layout_width="@dimen/dimen_24"
+            android:layout_height="@dimen/dimen_24"
+            app:srcCompat="@drawable/ic_support"
+            app:tint="@color/grey_500"
+            android:layout_gravity="end|center_vertical"
+            android:layout_marginEnd="16dp"/>
+    </FrameLayout>
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="@dimen/dimen_61">
+
+        <TextView
+            android:id="@+id/header_agreement"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textColor="@color/black"
+            android:textSize="@dimen/dimen_16s"
+            android:paddingStart="@dimen/dimen_16"
+            android:gravity="center_vertical"
+            android:layout_gravity="start|center_vertical"
+            android:text="@string/header_agreement"
+            tools:ignore="RtlSymmetry" />
+
+        <ImageView
+            android:id="@+id/button_agreement"
+            android:layout_width="@dimen/dimen_24"
+            android:layout_height="@dimen/dimen_24"
+            app:srcCompat="@drawable/ic_arrow_forward"
+            app:tint="@color/grey_500"
+            android:layout_gravity="end|center_vertical"
+            android:layout_marginEnd="@dimen/dimen_16"/>
+    </FrameLayout>
+</LinearLayout>
