+++ app/src/main/res/values-en/strings.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="header_one">Question:</string>
+    <string name="question">Is this movie rated more than 9?</string>
+    <string name="no">No</string>
+    <string name="yes">Yes</string>
+    <string name="search">Search</string>
+    <string name="media">Media library</string>
+    <string name="setting">Settings</string>
+    <string name="light_theme">Light theme</string>
+    <string name="dark_theme">Dark theme</string>
+    <string name="header_share">Share app</string>
+    <string name="header_support">Write to support</string>
+    <string name="header_agreement">Terms of use</string>
+</resources>
