+++ app/src/main/res/values/colors.xml
@@ -2,4 +2,8 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+    <color name="grey_100">#FF1A1B22</color>
+    <color name="grey_500">#FFAEAFB4</color>
+    <color name="grey_700">#FFE6E8EB</color>
+    <color name="background_one">#FF3772E7</color>
 </resources>
